Move files to presentation/common/compose
