+++ app/src/main/java/ru/internetcloud/strava/presentation/common/compose/Color.kt
@@ -1,4 +1,4 @@
-package ru.internetcloud.strava.presentation.main
+package ru.internetcloud.strava.presentation.common.compose
 
 import androidx.compose.ui.graphics.Color
