+++ app/src/main/java/ru/internetcloud/strava/presentation/training/detail/TrainingDetailScreen.kt
@@ -32,9 +32,9 @@ import ru.internetcloud.strava.R
 import ru.internetcloud.strava.domain.model.Profile
 import ru.internetcloud.strava.domain.model.Training
 import ru.internetcloud.strava.domain.util.TrainingConverter
-import ru.internetcloud.strava.presentation.main.ShowEmptyData
-import ru.internetcloud.strava.presentation.main.ShowError
-import ru.internetcloud.strava.presentation.main.ShowLoadingData
+import ru.internetcloud.strava.presentation.common.compose.ShowEmptyData
+import ru.internetcloud.strava.presentation.common.compose.ShowError
+import ru.internetcloud.strava.presentation.common.compose.ShowLoadingData
 import ru.internetcloud.strava.presentation.training.list.TimeDistanceSpeed
 import ru.internetcloud.strava.presentation.util.DateTimeConverter
 import ru.internetcloud.strava.presentation.util.UiState
