+++ app/src/main/java/ru/internetcloud/strava/presentation/training/list/TrainingListScreen.kt
@@ -19,9 +19,9 @@ import androidx.compose.ui.unit.dp
 import androidx.lifecycle.viewmodel.compose.viewModel
 import ru.internetcloud.strava.R
 import ru.internetcloud.strava.domain.model.ProfileWithTrainingList
-import ru.internetcloud.strava.presentation.main.ShowEmptyData
-import ru.internetcloud.strava.presentation.main.ShowError
-import ru.internetcloud.strava.presentation.main.ShowLoadingData
+import ru.internetcloud.strava.presentation.common.compose.ShowEmptyData
+import ru.internetcloud.strava.presentation.common.compose.ShowError
+import ru.internetcloud.strava.presentation.common.compose.ShowLoadingData
 import ru.internetcloud.strava.presentation.util.UiState
 
 @Composable
