+++ app/src/main/java/ru/internetcloud/strava/presentation/util/Utils.kt
@@ -1,0 +1,5 @@
+package ru.internetcloud.strava.presentation.util
+
+inline fun <T> T.applyIf(predicate: Boolean, block: T.() -> T): T {
+    return if (predicate) block() else this
+}
